viikko3/KumpulaBiershop/src/main/java/com/mycompany/webkauppa/ohjaus/OstoksenPoistoKorista.java

package com.mycompany.webkauppa.ohjaus;

import com.mycompany.webkauppa.sovelluslogiikka.Ostoskori;
import com.mycompany.webkauppa.sovelluslogiikka.Tuote;
import com.mycompany.webkauppa.sovelluslogiikka.Varasto;

public class OstoksenPoistoKorista {
    private Ostoskori ostoskori;
    private long tuoteId;
    private Varasto varasto;
    
    public OstoksenPoistoKorista(Ostoskori ostoskori, long tuoteId) {
        this.ostoskori = ostoskori;
        this.tuoteId = tuoteId;
        this.varasto = Varasto.getInstance();
    }    
    
    public void suorita() {
        varasto.palautaVarastoon( tuoteId );         
        Tuote poistettava = varasto.etsiTuote( tuoteId );              
        ostoskori.poista(poistettava);  
    }          
}
